Refactoriza variables de entorno
- Coloca lastConnection dentro de utils
- Agrupa opciones de conexion como variable env
- Elimina codigo duplicado
