+++ server.py
@@ -5,9 +5,6 @@
 from flask import Flask
 from flask_restx import Api
 
-# Modulos propios
-from utils.environment import Environment
-
 app = Flask(__name__)
 api = Api(
     app,
